Add class to compatibility with Magento 2.3
